+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 otj-server
 ==========
 
+3.0.7
+-----
+* Split Reactive Server in to its own top-level server component
+* Replace @RestReactiveHttpServer with @ReactiveServer annotation
+* Server request logging for Reactive Server
+* Reactive health and metrics endpoints for Reactive Server
+
 3.0.6
 -----
 * Fix EmbeddedReactiveJetty to work after module split (fixes regression post pom 180).